Vehicle and crate save/load system integration for Mike Force
The very early integration of my save/load script for vehicles and crates.

The area check function is non-operational which means everything will be saved and loaded which probably breaks a lot of decorations placed around the base and many functions are still filled with old comments and such but here it is in its early form.